refactor: redesign errors

+++ src/process/tracee.rs
@@ -27,7 +27,7 @@ impl TraceeStatus {
 
     pub fn get_errno(&self) -> i32 {
         match self {
-            TraceeStatus::Error(err) => err.get_errno(),
+            TraceeStatus::Error(err) => err.get_errno() as i32,
             _ => 0,
         }
     }
